Solution cleaning // Cleaning the solution after a break. + DAL.Model add ITable or IView + Minor changes in UI

+++ NorthwindWindowsStoreService/NorthwindWindowsStore.Service/BaseService.cs
@@ -1,13 +1,16 @@
 ﻿namespace NorthwindWindowsStore.Service
 {
+    using Nelibur.ObjectMapper;
     using NorthwindWindowsStore.DAL;
+    using NorthwindWindowsStore.DAL.Model.Interface;
+    using NorthwindWindowsStore.Service.Map;
     using System;
     using System.Collections.Generic;
     using System.Linq.Expressions;
     using System.Threading.Tasks;
 
     /// <summary>
-    /// 
+    /// Base service provided for services
     /// </summary>
     /// <typeparam name="I">Input - DAL model</typeparam>
     /// <typeparam name="O">Output - ViewModel</typeparam>
@@ -15,16 +18,34 @@
         where I : class, new()
         where O : class, new()
     {
-        private BaseRepository<I> _repository;
+        private BaseRepository<I> _repository { get; set; }
+        private BaseView<I> _view { get; set; }
+        private static bool MapperIsntSet = true;
 
         public BaseService(String applicationRunDirectory)
         {
-            _repository = new BaseRepository<I>(applicationRunDirectory);
+            List<Type> interfaces = new List<Type>();
+            interfaces.AddRange(typeof(I).GetInterfaces() as Type[]);
+
+            if (interfaces.Contains(typeof(ITable)))
+            {
+                _repository = new BaseRepository<I>(applicationRunDirectory);
+            }
+            else if (interfaces.Contains(typeof(IView)))
+            {
+                _view = new BaseView<I>(applicationRunDirectory);
+            }
+
+            if (MapperIsntSet)
+            {
+                MapperIsntSet = false;
+                TinyMapperStaticConfiguration.SetTinyMapperStaticConfiguration();
+            }
         }
 
         protected O TransformObject(I input)
         {
-            return null;
+            return TinyMapper.Map<O>(input);
         }
 
         protected IEnumerable<O> TranformList(IEnumerable<I> inputList)
@@ -35,19 +56,31 @@
             }
         }
 
-        public async void Add(I item)
+        public async Task<bool> Add(I item)
         {
-            await _repository.Add(item);
+            if (_repository != null)
+            {
+                return false;
+            }
+            return await _repository.Add(item);
         }
 
-        public async void Remove(I item)
+        public async Task<bool> Remove(I item)
         {
-            _repository.Remove(item);
+            if (_repository == null)
+            {
+                return false;
+            }
+            return await _repository.Remove(item);
         }
 
-        public async void Update(I item)
+        public async Task<bool> Update(I item)
         {
-            _repository.Update(item);
+            if (_repository == null)
+            {
+                return false;
+            }
+            return await _repository.Update(item);
         }
 
         public async Task<O> FindByID(int id)
@@ -59,9 +92,15 @@
             return TranformList(_repository.Find(predicate).Result);
         }
 
-        public async Task<IEnumerable<O>> GetAll()
+        public IEnumerable<O> GetAll()
         {
-            return TranformList(_repository.GetAll().Result);
+            var result = _repository.GetAll();
+            return TranformList(result);
+        }
+
+        public async Task<IEnumerable<O>> GetViewResult()
+        {
+            return TranformList(_view.GetViewResult().Result);
         }
     }
 }
